+++ src/lib/sh68f90a/clock.h
