+++ src/lib/sh68f90a/usbregs.h
