+++ src/matrix.c
@@ -1,7 +1,7 @@
 #include "matrix.h"
-#include "sh68f90a.h"
-#include "pwm.h"
-#include "delay.h"
+#include "lib/sh68f90a/sh68f90a.h"
+#include "lib/sh68f90a/pwm.h"
+#include "lib/sh68f90a/delay.h"
 #include "layout.h"
 #include "report.h"
 #include "debug.h"
